Add support to ignore certain ldap search requests.
BMCs, Dell idracs especially seem to lookup a system user in LDAP,
e.g "Administrator" this user is locally configured on the BMC,
even then its looked up on the ldap server.

This results in unnecessary requests to bmcldap,
to prevent this we allow a slice of strings to be declared in the config,
if any of these strings are found in a LDAP search request,
that search is not performed on the backend LDAP server,
instead a ldap.ResultUnwillingToPerform is returned to the client.

IgnoreFilters:
  - "uid=foobar"
  - "Administrator"